#58 - Rename main module to prosimos, expose referenced modules in pypi package
