+++ prosimos/prioritisation_parser.py
@@ -1,5 +1,5 @@
 from typing import List
-from bpdfr_simulation_engine.prioritisation_rules import (
+from prosimos.prioritisation_rules import (
     AllPriorityRules,
     AndPrioritisationRule,
     OrPrioritisationRule,
