+++ testing_scripts/bpm22_experiments_script.py
@@ -5,8 +5,8 @@ import os
 
 from bpdfr_discovery.log_parser import discover_aggregated_task_distributions, preprocess_xes_log, \
     discover_resource_task_duration_distribution, save_prosimos_json
-from bpdfr_simulation_engine.probability_distributions import create_default_distribution
-from bpdfr_simulation_engine.resource_calendar import build_full_time_calendar, CalendarFactory
+from prosimos.probability_distributions import create_default_distribution
+from prosimos.resource_calendar import build_full_time_calendar, CalendarFactory
 from testing_scripts.best_parameters_extraction import compute_median_simulation_emd
 
 from testing_scripts.bpm_2022_testing_files import experiment_logs, process_files, out_folder, canonical_json
