+++ testing_scripts/test_batching.py
@@ -6,7 +6,7 @@ from pathlib import Path
 import pytest
 import json
 from pandas import testing as tm
-from bpdfr_simulation_engine.simulation_properties_parser import BATCH_PROCESSING_SECTION
+from prosimos.simulation_properties_parser import BATCH_PROCESSING_SECTION
 
 from testing_scripts.bimp_diff_sim_tests import run_diff_res_simulation
 from testing_scripts.test_simulation import _verify_activity_count_and_duration
