+++ testing_scripts/test_batching_ready_wt.py
@@ -1,8 +1,8 @@
 import pytest
 import pandas as pd
 
-from bpdfr_simulation_engine.batch_processing import AndFiringRule, FiringSubRule, OrFiringRule
-from bpdfr_simulation_engine.resource_calendar import parse_datetime
+from prosimos.batch_processing import AndFiringRule, FiringSubRule, OrFiringRule
+from prosimos.resource_calendar import parse_datetime
 from testing_scripts.test_batching import (
     _verify_logs_ordered_asc,
     _verify_same_resource_for_batch,
