+++ testing_scripts/test_case_priority_is_true.py
@@ -1,5 +1,5 @@
 import pytest
-from bpdfr_simulation_engine.prioritisation_rules import (
+from prosimos.prioritisation_rules import (
     AndPrioritisationRule,
     OrPrioritisationRule,
     PrioritisationRule,
