+++ testing_scripts/test_case_priority_simulation.py
@@ -2,7 +2,7 @@ import json
 
 import pandas as pd
 import pytest
-from bpdfr_simulation_engine.simulation_properties_parser import (
+from prosimos.simulation_properties_parser import (
     BATCH_PROCESSING_SECTION,
     CASE_ATTRIBUTES_SECTION,
     PRIORITISATION_RULES_SECTION,
