+++ testing_scripts/test_is_enabled.py
@@ -1,11 +1,11 @@
 import datetime
 
 import pytz
-from bpdfr_simulation_engine.simulation_properties_parser import (
+from prosimos.simulation_properties_parser import (
     parse_json_sim_parameters,
     parse_simulation_model,
 )
-from bpdfr_simulation_engine.simulation_setup import SimDiffSetup
+from prosimos.simulation_setup import SimDiffSetup
 
 from test_discovery import assets_path
